Restructured file structure. Added negTimes.log as a test.
