+++ src/interfaces/LogstfInterfaces.ts
