+++ src/modules/PlayerStatsModule.ts
@@ -1,4 +1,4 @@
-import * as events from '../events'
+import * as events from '../interfaces/events'
 import { IGameState, PlayerInfo } from '../Game'
 import { renameObjectKeys } from "../Utilities"
 import { ClassStat, Medicstats, Player, Players, WeaponStats } from "../interfaces/LogstfInterfaces"
